fix: clean up link extraction
